New, Improved CEGUI theme: LineSpace
The CEGUI theme has diverged enough that it needs a new name. The new
theme, LineSpace, is a mashup of the SleekSpace and TaharezLook themes.
Earlier incarnations of the SleekSpace theme (designed for the game
Precursors) were closer to my desired look-and-feel. However, SleekSpace
left several UI widgets undefined or partially defined.

LineSpace is intended to be a drop-in replacement for TaharezLook (which
makes it easier for newbies to get up and running). If I require other
widgets not defined by TaharezLook, then LineSpace will become a
superset of the TaharezLook theme.

The LineSpace looknfeel file has been broken into multiple files (as on
the Precursors project). Thanks again to the Precursors team for their
creation of the SleekSpace theme - they made it a lot easier for me to
get to where I wanted to be:
David H. Bronke and Christopher S. Case
http://dev.skewedaspect.com/projects/precursors

A test CEGUI window has been created to test all widgets: TestWindow

Timesteps are now injected into the CEGUI system.

Known bugs:
* LineSpace imageset elements don't all line up properly when rendered.
* Several required imageset elements are blank.
* Tooltips are not appearing when hovering over widgets.
* The images and looknfeel still require cleanup.